requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
keyboard==0.13.5
MarkupSafe==3.0.3
MouseInfo==0.1.3
mss==10.1.0
pillow==12.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==2.0.1
pyobjc-core==12.1
pyobjc-framework-Cocoa==12.1
pyperclip==1.11.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
rubicon-objc==0.5.2
Werkzeug==3.1.3
